+++ Tests/Integration/xFileUpload.config.ps1
@@ -1,0 +1,30 @@
+param
+(
+    [Parameter(Mandatory = $true)]
+    [String]
+    $ConfigurationName
+)
+
+Configuration $ConfigurationName
+{
+    param
+    (
+        [Parameter(Mandatory = $true)]
+        [ValidateNotNullOrEmpty()]
+        [System.String]
+        $DestinationPath,
+
+        [Parameter(Mandatory = $true)]
+        [ValidateNotNullOrEmpty()]
+        [System.String]
+        $SourcePath
+    )
+
+    Import-DscResource -ModuleName 'xPSDesiredStateConfiguration'
+
+    xFileUpload UploadFileOrFolder
+    {
+        DestinationPath = $DestinationPath
+        SourcePath = $SourcePath
+    }
+}